ts to js hotfix
